Fixed constraint bug; replace constraint arrays with direct aplication of constraints; remove unnecessary whitespaces
