+++ ComicsApp/ComicsApp/Pokedex/Networking/PokemonApiCaller.swift
@@ -38,5 +38,5 @@ public class PokemonApiCaller {
             }
         }.resume()
     }
-        
+    
 }
